refactor: :recycle: home page query 분리
- #114
