+++ app/src/app.module.ts
@@ -3,8 +3,8 @@ import { Module } from '@nestjs/common';
 import { GraphQLModule } from '@nestjs/graphql';
 import { MongooseModule } from '@nestjs/mongoose';
 import { join } from 'path';
+import { CursusUserModule } from './api/cursusUser/cursusUser.module';
 import { ProjectModule } from './api/project/project.module';
-import { ScaleTeamModule } from './api/scaleTeam/scaleTeam.module';
 import { EvalLogModule } from './page/evalLog/evalLog.module';
 import { HomeModule } from './page/home/home.module';
 import { LeaderboardModule } from './page/leaderboard/leaderboard.module';
@@ -17,14 +17,14 @@ import { ProjectInfoModule } from './page/projectInfo/projectInfo.module';
     MongooseModule.forRoot(
       `mongodb://${process.env.DB_USERNAME}:${process.env.DB_PASSWORD}@${process.env.DB_ENDPOINT}/${process.env.DB_NAME}`,
     ),
+    CursusUserModule,
     ProjectModule,
     HomeModule,
+    LeaderboardModule,
+    EvalLogModule,
     ProjectInfoModule,
     PersonalGeneralModule,
     PersonalEvalModule,
-    LeaderboardModule,
-    EvalLogModule,
-    ScaleTeamModule,
     GraphQLModule.forRoot<ApolloDriverConfig>({
       driver: ApolloDriver,
       buildSchemaOptions: {
