+++ app/src/common/models/common.rate.model.ts
@@ -1,7 +1,7 @@
 import { Field, ObjectType } from '@nestjs/graphql';
 
 @ObjectType()
-export class NumericRate {
+export class IntRate {
   @Field()
   total: number;
