+++ app/src/page/home/eval/models/home.eval.model.ts
@@ -1,0 +1,23 @@
+import { Field, ObjectType } from '@nestjs/graphql';
+import {
+  FloatDateRanged,
+  IntDateRanged,
+} from 'src/common/models/common.dateRanaged.model';
+
+@ObjectType()
+export class HomeEval {
+  @Field()
+  totalEvalCount: number;
+
+  @Field()
+  evalCountByDateTemplate: IntDateRanged;
+
+  @Field()
+  averageEvalCountByDateTemplate: FloatDateRanged;
+
+  @Field()
+  averageFeedbackLength: number;
+
+  @Field()
+  averageCommentLength: number;
+}
